feat: add card layout to a11y status component
