+++ package.json
@@ -57,7 +57,7 @@
     "gatsby": "^4.25.7",
     "gatsby-image": "^3.7.1",
     "gatsby-plugin-image": "^2.9.0",
-    "gatsby-theme-carbon": "^3.2.3",
+    "gatsby-theme-carbon": "^3.4.0",
     "lodash-es": "^4.17.15",
     "markdown-it": "^12.3.2",
     "nanoid": "^2.1.11",
