+++ src/components/A11yStatus/a11y-status.module.scss
@@ -21,3 +21,7 @@
 :global(.cds--tag .cds--popover) {
   white-space: normal;
 }
+
+.variant + .card-group {
+  --space: 0;
+}
