+++ src/components/A11yStatus/index.js
@@ -1,0 +1,3 @@
+import A11yStatus from './A11yStatus';
+
+export default A11yStatus;
