--- src/components/A11yStatusTable/index.js
@@ -1,3 +1,0 @@
-import A11yStatusTable from './A11yStatusTable';
-
-export default A11yStatusTable;
