+++ src/pages/components/UI-shell-left-panel/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -34,7 +34,7 @@ component.
 
 </AnchorLinks>
 
-<A11yStatusTable components="UI shell" />
+<A11yStatus layout="table" components="UI shell" />
 
 ## What Carbon provides
