+++ src/pages/components/accordion/accessibility.mdx
@@ -7,7 +7,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <PageDescription>
 
@@ -22,7 +22,7 @@ standard accordion component, Carbon already incorporates accessibility.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Accordion" />
+<A11yStatus layout="table" components="Accordion" />
 
 ## What Carbon provides
