+++ src/pages/components/button/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <PageDescription>
 
@@ -21,7 +21,7 @@ standard button component, Carbon already incorporates accessibility.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Button" />
+<A11yStatus layout="table" components="Button" />
 
 ## What Carbon provides
