+++ src/pages/components/code-snippet/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -32,7 +32,7 @@ component.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Code snippet" />
+<A11yStatus layout="table" components="Code snippet" />
 
 ## What Carbon provides
