+++ src/pages/components/content-switcher/accessibility.mdx
@@ -7,7 +7,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -33,7 +33,7 @@ standard content switcher component, Carbon already incorporates accessibility.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Content switcher" />
+<A11yStatus layout="table" components="Content switcher" />
 
 ## What Carbon provides
