+++ src/pages/components/data-table/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <PageDescription>
 
@@ -21,7 +21,7 @@ standard data table component, Carbon already incorporates accessibility.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Data table" />
+<A11yStatus layout="table" components="Data table" />
 
 ## What Carbon provides
