+++ src/pages/components/dropdown/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <PageDescription>
 
@@ -23,7 +23,8 @@ component.
 
 </AnchorLinks>
 
-<A11yStatusTable
+<A11yStatus
+  layout="table"
   components={[
     'Dropdown',
     'Multiselect',
