+++ src/pages/components/dropdown/usage.mdx
@@ -7,6 +7,12 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
+import A11yStatus from 'components/A11yStatus';
+
+<A11yStatus layout="cards" components={['Dropdown', 'Multiselect']} />
+
+<A11yStatus layout="cards" components="Button" />
+
 <PageDescription>
 
 Dropdowns present a list of options from which a user can select one option, or
