+++ src/pages/components/form/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -34,7 +34,8 @@ considerations.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable
+<A11yStatus
+  layout="table"
   components={['Form', 'Form group', 'Form label', 'Fluid form']}
 />
