+++ src/pages/components/inline-loading/accessibility.mdx
@@ -3,7 +3,7 @@ title: Inline loading
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -31,7 +31,7 @@ component.
 
 </AnchorLinks>
 
-<A11yStatusTable components="Inline loading" />
+<A11yStatus layout="table" components="Inline loading" />
 
 ## What Carbon provides
