+++ src/pages/components/list/accessibility.mdx
@@ -5,7 +5,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -23,7 +23,7 @@ import {
   <AnchorLink>Developer considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components={['Ordered list', 'Unordered list']} />
+<A11yStatus layout="table" components={['Ordered list', 'Unordered list']} />
 
 ## What Carbon provides
