+++ src/pages/components/loading/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -34,7 +34,7 @@ standard loading component, Carbon already incorporates accessibility.
 
 </AnchorLinks>
 
-<A11yStatusTable components="Loading" />
+<A11yStatus layout="table" components="Loading" />
 
 ## What Carbon provides
