+++ src/pages/components/menu-buttons/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -41,7 +41,7 @@ button is going to be updated soon.
 
 </AnchorLinks>
 
-<A11yStatusTable components="Menu buttons" />
+<A11yStatus layout="table" components="Menu buttons" />
 
 ## What Carbon provides
