+++ src/pages/components/menu/accessibility.mdx
@@ -4,7 +4,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <InlineNotification>
 
@@ -12,4 +12,4 @@ import A11yStatusTable from 'components/A11yStatusTable';
 
 </InlineNotification>
 
-<A11yStatusTable components="Menu" />
+<A11yStatus layout="table" components="Menu" />
