+++ src/pages/components/modal/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -32,7 +32,7 @@ standard modal dialog component, Carbon already incorporates accessibility.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components={['Modal', 'Composed modal']} />
+<A11yStatus layout="table" components={['Modal', 'Composed modal']} />
 
 ## What Carbon provides
