+++ src/pages/components/notification/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -26,7 +26,7 @@ import {
   <AnchorLink>Accessibility testing</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Notification" />
+<A11yStatus layout="table" components="Notification" />
 
 ## How it works
