+++ src/pages/components/number-input/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -34,7 +34,10 @@ standard number input component, Carbon already incorporates accessibility.
 
 </AnchorLinks>
 
-<A11yStatusTable components={['Number input', 'Fluid number input']} />
+<A11yStatus
+  layout="table"
+  components={['Number input', 'Fluid number input']}
+/>
 
 ## What Carbon provides
