+++ src/pages/components/overview/accessibility-status.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Components', 'Accessibility status']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 import A11yStatusTag from 'components/A11yStatusTag';
 
 <PageDescription>
@@ -59,4 +59,4 @@ been tested for accessibility.
 | Keyboard navigation | Tests that ensure focus is properly managed, and all interactive functions of a component have a proper keyboard-accessible equivalent.                            | <A11yStatusTag tag="tested" /> <br/><A11yStatusTag tag="notavailable" /><br/><A11yStatusTag tag="partial" />                                       |
 | Screen reader       | This manual testing ensures that the visual information on the screen is properly conveyed and read correctly by screen readers such as JAWS, VoiceOver, and NVDA. | <A11yStatusTag tag="notavailable" /><br/><A11yStatusTag tag="nottested" /><br/><A11yStatusTag tag="manual" /> <br/><A11yStatusTag tag="partial" /> |
 
-<A11yStatusTable />
+<A11yStatus />
