+++ src/pages/components/progress-bar/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <PageDescription>
 
@@ -23,7 +23,7 @@ standard progress bar component, Carbon already incorporates accessibility.
 
 </AnchorLinks>
 
-<A11yStatusTable components="Progress bar" />
+<A11yStatus layout="table" components="Progress bar" />
 
 ## What Carbon provides
