+++ src/pages/components/progress-indicator/accessibility.mdx
@@ -7,7 +7,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -35,7 +35,7 @@ be updated when these accessibility issues are resolved.
   <AnchorLink>Accessibility testing</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Progress indicator" />
+<A11yStatus layout="table" components="Progress indicator" />
 
 ## Accessibility considerations
