+++ src/pages/components/search/accessibility.mdx
@@ -3,7 +3,7 @@ title: Search
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -32,7 +32,7 @@ annotations are needed to make the search messaging and results accessible.
 
 </AnchorLinks>
 
-<A11yStatusTable components={['Search', 'Fluid search']} />
+<A11yStatus layout="table" components={['Search', 'Fluid search']} />
 
 ## What Carbon provides
