+++ src/pages/components/select/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -34,7 +34,7 @@ component.
 
 </AnchorLinks>
 
-<A11yStatusTable components={['Select', 'Fluid select']} />
+<A11yStatus layout="table" components={['Select', 'Fluid select']} />
 
 ## What Carbon provides
