+++ src/pages/components/slider/accessibility.mdx
@@ -7,7 +7,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import localVideoOne from './videos/default-slider-demo.mp4';
 import localPosterOne from './images/slider-accessibility-default-mp4-image.png';
@@ -27,7 +27,7 @@ component.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Slider" />
+<A11yStatus layout="table" components="Slider" />
 
 ## What Carbon provides
