+++ src/pages/components/structured-list/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -34,7 +34,7 @@ be updated when these accessibility issues are resolved.
   <AnchorLink>Accessibility testing</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Structured list" />
+<A11yStatus layout="table" components="Structured list" />
 
 ## Accessibility considerations
