+++ src/pages/components/tabs/accessibility.mdx
@@ -5,7 +5,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -32,7 +32,7 @@ accessibility.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Tabs" />
+<A11yStatus layout="table" components="Tabs" />
 
 ## What Carbon provides
