+++ src/pages/components/tabs/code.mdx
@@ -5,7 +5,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import { Tabs, TabList, Tab, TabPanels, TabPanel } from '@carbon/react';
