+++ src/pages/components/toggle/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import {
   StructuredListWrapper,
@@ -35,7 +35,7 @@ component.
 
 </AnchorLinks>
 
-<A11yStatusTable components="Toggle" />
+<A11yStatus layout="table" components="Toggle" />
 
 ## What Carbon provides
