+++ src/pages/components/toggletip/accessibility.mdx
@@ -7,7 +7,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 import { Toggletip } from '@carbon/react';
 
@@ -23,7 +23,7 @@ trigger element and can contain interactive elements.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Toggletip" />
+<A11yStatus layout="table" components="Toggletip" />
 
 ## What Carbon provides
