+++ src/pages/components/tooltip/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <PageDescription>
 
@@ -23,7 +23,7 @@ annotate the tooltip’s text content.
 
 </AnchorLinks>
 
-<A11yStatusTable components="Tooltip" />
+<A11yStatus layout="table" components="Tooltip" />
 
 ## What Carbon provides
