+++ src/pages/components/tree-view/accessibility.mdx
@@ -6,7 +6,7 @@ description:
 tabs: ['Usage', 'Style', 'Code', 'Accessibility']
 ---
 
-import A11yStatusTable from 'components/A11yStatusTable';
+import A11yStatus from 'components/A11yStatus';
 
 <PageDescription>
 
@@ -21,7 +21,7 @@ component.
   <AnchorLink>Development considerations</AnchorLink>
 </AnchorLinks>
 
-<A11yStatusTable components="Treeview" />
+<A11yStatus layout="table" components="Treeview" />
 
 ## What Carbon provides
